requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
asttokens==2.4.1
async_timeout==5.0.1
attrs==25.3.0
be_great==0.0.8
cachetools==5.5.0
certifi==2024.8.30
charset_normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dask==2024.11.2
datasets==3.5.0
debugpy==1.8.5
decorator==5.1.1
dill==0.3.8
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.18.0
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.10.0
huggingface_hub==0.30.2
idna==3.10
imbalanced_learn==0.13.0
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.26.0
japanize_matplotlib==1.1.3
jedi==0.19.1
jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kanjize==1.5.0
kiwisolver==1.4.5
lightgbm==4.6.0
locket==1.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib_inline==0.1.7
mord==0.7
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
nest_asyncio==1.6.0
networkx==3.4.2
normalize_japanese_addresses==0.0.9
numpy==2.1.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
partd==1.4.2
patsy==1.0.1
pillow==10.4.0
pip==24.3.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
propcache==0.3.1
psutil==6.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pygments==2.18.0
pyparsing==3.1.4
python_dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==65.5.0
six==1.16.0
sklearn_compat==0.1.3
stack_data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
toolz==1.0.0
torch==2.6.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.51.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
shap